+++ guillotina/api/__init__.py
@@ -6,6 +6,7 @@ from . import container  # noqa
 from . import content  # noqa
 from . import dynamic  # noqa
 from . import files  # noqa
+from . import login  # noqa
 from . import registry  # noqa
 from . import search  # noqa
 from . import storage  # noqa